agent-backend/pyproject.toml
[project]
name = "kenya-constitution-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "greenlet>=3.3.2",
    "loguru>=0.7.3",
    "openai>=2.30.0",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.77.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
